#23405 remove unused code
