+++ dotCMS/src/test/java/com/dotmarketing/util/ZipUtilTest.java
@@ -4,8 +4,11 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
 import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.nio.charset.Charset;
 import java.nio.file.Files;
 import java.util.List;
 import java.util.zip.ZipEntry;
@@ -41,6 +44,54 @@ public class ZipUtilTest  {
           }
      }
 
+     /**
+      * Simply test that we cn only unzip under the properly configured out dir
+      * Here we expect success
+      * @throws IOException
+      */
+     @Test
+     public void testUnzipUtilSecurityCheckHappyScenario() throws IOException {
+          final long time = java.lang.System.nanoTime();
+          final File zipEntry = File.createTempFile("zip-entry-" + time, ".txt");
+          FileUtils.writeStringToFile(zipEntry, "" + System.currentTimeMillis(),
+                  Charset.defaultCharset());
+          final File zip = File.createTempFile("zip-test" + time, ".zip");
+          makeZip(zipEntry, zip.getCanonicalPath());
+          final String destinationDir = ConfigUtils.getIntegrityPath();
+          try (
+                  final FileInputStream inputStream = new FileInputStream(zip);
+          ) {
+               ZipUtil.extract(inputStream, destinationDir);
+          }
+     }
+
+     /**
+      * Good zip
+      * @param file
+      * @param zipFileName
+      * @throws IOException
+      */
+     private static void makeZip(final File file, final String zipFileName) throws IOException {
+
+          try (
+                  //create ZipOutputStream to write to the zip file
+                  FileOutputStream fos = new FileOutputStream(zipFileName);
+                  ZipOutputStream zos = new ZipOutputStream(fos);
+          ) {
+               //add a new Zip Entry to the ZipOutputStream
+               ZipEntry ze = new ZipEntry(file.getName());
+               zos.putNextEntry(ze);
+               //read the file and write to ZipOutputStream
+
+               try (FileInputStream fis = new FileInputStream(file);) {
+                    byte[] buffer = new byte[1024];
+                    int len;
+                    while ((len = fis.read(buffer)) > 0) {
+                         zos.write(buffer, 0, len);
+                    }
+               }
+          }
+     }
 
      /**
       * Creates a zip file that includes a file that attempts a directory traversal
